requirements.txt
absl-py==1.0.0
astunparse==1.6.3
attrs==21.4.0
autopy==4.0.0
cachetools==5.0.0
certifi==2021.10.8
charset-normalizer==2.0.12
cvzone==1.5.6
cycler==0.11.0
flatbuffers==2.0
fonttools==4.32.0
gast==0.5.3
google-auth==2.6.4
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.44.0
h5py==3.6.0
idna==3.3
importlib-metadata==4.11.3
keras==2.8.0
Keras-Preprocessing==1.1.2
kiwisolver==1.4.2
libclang==13.0.0
Markdown==3.3.6
matplotlib==3.5.1
mediapipe==0.8.9.1
numpy==1.22.3
oauthlib==3.2.0
opencv-contrib-python==4.5.5.64
opencv-python==4.5.5.64
opt-einsum==3.3.0
packaging==21.3
Pillow==9.1.0
protobuf==3.20.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==3.0.8
python-dateutil==2.8.2
requests==2.27.1
requests-oauthlib==1.3.1
rsa==4.8
six==1.16.0
tensorboard==2.8.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.8.0
tensorflow-io-gcs-filesystem==0.24.0
termcolor==1.1.0
tf-estimator-nightly==2.8.0.dev2021122109
typing_extensions==4.1.1
urllib3==1.26.9
Werkzeug==2.1.1
wrapt==1.14.0
zipp==3.8.0
